[3722] Fix reference widget registration
Bug: https://github.com/eclipse-sirius/sirius-web/issues/3722
