+++ packages/sirius-web/backend/sirius-web-infrastructure/src/main/java/org/eclipse/sirius/web/infrastructure/emf/EMFChildExtenderProviderConfiguration.java
@@ -18,8 +18,10 @@ import org.eclipse.sirius.components.view.deck.provider.DeckItemProviderAdapterF
 import org.eclipse.sirius.components.view.diagram.DiagramPackage;
 import org.eclipse.sirius.components.view.diagram.customnodes.provider.CustomnodesItemProviderAdapterFactory;
 import org.eclipse.sirius.components.view.diagram.provider.DiagramItemProviderAdapterFactory;
+import org.eclipse.sirius.components.view.form.FormPackage;
 import org.eclipse.sirius.components.view.form.provider.FormItemProviderAdapterFactory;
 import org.eclipse.sirius.components.view.gantt.provider.GanttItemProviderAdapterFactory;
+import org.eclipse.sirius.components.widgets.reference.provider.ReferenceItemProviderAdapterFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -54,5 +56,10 @@ public class EMFChildExtenderProviderConfiguration {
     @Bean
     public ChildExtenderProvider customNodesChildExtenderProvider() {
         return new ChildExtenderProvider(DiagramPackage.eNS_URI, CustomnodesItemProviderAdapterFactory.DiagramChildCreationExtender::new);
+    }
+
+    @Bean
+    public ChildExtenderProvider referenceWidgetChildExtenderProvider() {
+        return new ChildExtenderProvider(FormPackage.eNS_URI, ReferenceItemProviderAdapterFactory.FormChildCreationExtender::new);
     }
 }
